documentation updated
