+++ rest-assured/src/main/java/co/com/sofka/model/PostModel.java
@@ -1,5 +1,10 @@
 package co.com.sofka.model;
 
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
 public class PostModel {
     private String id;
     private String title;
@@ -7,45 +12,6 @@ public class PostModel {
     private String userId;
 
     public PostModel() {
-    }
-
-    public PostModel(String id, String title, String body, String userId) {
-        this.id = id;
-        this.title = title;
-        this.body = body;
-        this.userId = userId;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getBody() {
-        return body;
-    }
-
-    public void setBody(String body) {
-        this.body = body;
-    }
-
-    public String getUserId() {
-        return userId;
-    }
-
-    public void setUserId(String userId) {
-        this.userId = userId;
     }
 
     @Override
