+++ rest-assured/src/main/java/co/com/sofka/model/reqress/UserGetModel.java
@@ -1,0 +1,26 @@
+package co.com.sofka.model.reqress;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+public class UserGetModel {
+    private String id;
+    private String email;
+    private String firstName;
+    private String lastName;
+
+    public UserGetModel() {
+    }
+
+    @Override
+    public String toString() {
+        return "UserModel{" +
+                "id='" + id + '\'' +
+                ", email='" + email + '\'' +
+                ", firstName='" + firstName + '\'' +
+                ", lastName='" + lastName + '\'' +
+                '}';
+    }
+}
