+++ rest-assured/src/test/java/co/com/sofka/stepdefinitions/CreateUserStepDefinition.java
@@ -1,6 +1,7 @@
 package co.com.sofka.stepdefinitions;
 
-import co.com.sofka.model.UserPostModel;
+import co.com.sofka.model.reqress.UserPostModel;
+import co.com.sofka.model.reqress.UserPostResponseModel;
 import co.com.sofka.util.SetModelData;
 import com.google.gson.Gson;
 import io.cucumber.datatable.DataTable;
@@ -15,9 +16,11 @@ import java.util.Map;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.notNullValue;
+import static org.hamcrest.MatcherAssert.assertThat;
 
 public class CreateUserStepDefinition extends Setup {
     private UserPostModel userPostModel;
+    private final Gson gson = new Gson();
 
 
     @Given("the user is currently on the reqres.in platform.")
@@ -32,9 +35,8 @@ public class CreateUserStepDefinition extends Setup {
 
         List<Map<String, String>> userDataList = dataTable.asMaps(String.class, String.class);
         userPostModel = SetModelData.mapUserPostModelData(userDataList);
-
-        Gson gson = new Gson();
         String json = gson.toJson(userPostModel);
+
 
         request.header("Content-Type", "application/json");
         request.body(json);
@@ -43,9 +45,16 @@ public class CreateUserStepDefinition extends Setup {
 
     @Then("they should be able to view their information along with a generated ID")
     public void theyShouldBeAbleToViewTheirInformationAlongWithAGeneratedID() {
-        response.then().assertThat().body("id", notNullValue());
-        response.then().assertThat().body("name", equalTo(userPostModel.getName()));
-        response.then().assertThat().body("job", equalTo(userPostModel.getJob()));
+
+        // Deserialization from json
+        String jsonResponseString = response.asString();
+        UserPostResponseModel userResponse = new Gson().fromJson(jsonResponseString,
+                UserPostResponseModel.class);
+
+        // Compare response model with sent model
+        assertThat(userResponse.getId(), notNullValue());
+        assertThat(userResponse.getName(), equalTo(userPostModel.getName()));
+        assertThat(userResponse.getJob(), equalTo(userPostModel.getJob()));
     }
 
 }
