+++ rest-assured/src/test/resources/features/create_user.feature
@@ -14,7 +14,10 @@ Feature: Account Creation on reqres.in Platform
     Then they should be able to view their information along with a generated ID
     And the response should have a valid status code of 201
     Examples:
-      | name            | job   |
-      | johan@gmail.com | johan |
-      | jorge@gmail.com | jorge |
+      | name                | job       |
+      | johan@gmail.com     | johan     |
+      | jorge@gmail.com     | jorge     |
+      | juan_jose@gmail.com | juan_jose |
+      | rojas@gmail.com     | rojas     |
+
